requirements.txt
openai
twilio
python-dotenv
requests
